Move signals/events code to own file.
