+++ src/Makefile
@@ -43,6 +43,7 @@ OBJS=	mosquitto.o \
 		send_subscribe.o \
 		send_unsubscribe.o \
 		service.o \
+		signals.o \
 		subs.o \
 		sys_tree.o \
 		time_mosq.o \
@@ -158,6 +159,9 @@ send_unsubscribe.o : ../lib/send_unsubscribe.c ../lib/send_mosq.h
 	${CROSS_COMPILE}${CC} $(BROKER_CFLAGS) -c $< -o $@
 
 service.o : service.c mosquitto_broker.h
+	${CROSS_COMPILE}${CC} $(BROKER_CFLAGS) -c $< -o $@
+
+signals.o : signals.c mosquitto_broker.h
 	${CROSS_COMPILE}${CC} $(BROKER_CFLAGS) -c $< -o $@
 
 subs.o : subs.c mosquitto_broker.h
